move inner class CategoryCsvLineValue to top level
